add GH action and clean authent logic
