+++ components/immatriculations/rncs.tsx
@@ -8,6 +8,7 @@ import { IImmatriculationRNCS } from '../../models/immatriculation';
 import { inpi } from '../../public/static/logo';
 import AdministrationNotResponding from '../administration-not-responding';
 import ButtonLink from '../button';
+import { download } from '../icon';
 import { Section } from '../section';
 
 interface IProps {
@@ -30,14 +31,24 @@ const ImmatriculationRNCS: React.FC<IProps> = ({ immatriculation }) => {
         >
           <div className="description">
             <div>
-              Cette entité possède une fiche d'immatriculation sur le{' '}
+              Cette entité possède une fiche d'immatriculation au{' '}
               <b>Registre National du Commerce et des Sociétés (RNCS)</b> qui
               liste les entreprises enregistrées auprès des Greffes des
               tribunaux de commerce et centralisées par l'INPI.
+              <p>
+                Pour accéder à l’ensemble des données contenues dans un extrait
+                KBIS, vous pouvez télécharger le justificatif d’immatriculation.
+                Si le téléchargement échoue, vous pouvez accéder à la donnée en
+                allant sur le site de l'INPI. Pour accéder à la donnée complète,
+                vous <b>devrez vous créer un compte INPI</b>.
+              </p>
             </div>
             <div className="logo-wrapper">{inpi}</div>
           </div>
           <div className="layout-center">
+            <ButtonLink href={`/api/inpi-pdf-proxy/${immatriculation.siren}`}>
+              {download} Télécharger le justificatif d'immatriculation
+            </ButtonLink>
             <div className="separator" />
             <ButtonLink
               target="_blank"
