+++ pages/donnees-extrait-kbis.tsx
@@ -138,7 +138,7 @@ const ExtraitKbis: React.FC = () => {
           <p>
             Il existe une API spéciale, <b>API Entreprise</b> réservée aux
             administrations et à certaines entreprises éligibles, qui permet
-            d'accèder aux informations des entreprises détenue par
+            d'accéder aux informations des entreprises détenue par
             l'administration :
           </p>
           <ul>
